Path `sklearn.utils.discovery` as well
Some tests import `all_estimators` from the full path instead of
`sklearn.utils`. We need to patch both locations.
